requirements.txt
-i https://pypi.org/simple
-e ./rpi-rgb-led-matrix/bindings/python
certifi==2020.6.20
chardet==3.0.4
dbus-python==1.2.16
idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
importlib-metadata==1.7.0; python_version < '3.8'
inflect==4.1.0
numpy==1.19.1; python_version >= '3.6'
opencv-python-headless==4.3.0.36
pillow==7.2.0
python-dotenv==0.14.0
requests-mock==1.8.0
requests==2.24.0
six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
urllib3==1.25.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
zipp==3.1.0; python_version >= '3.6'
